Restyled by yapf

+++ tensorflow/python/ops/nn_fused_batchnorm_test.py
@@ -41,11 +41,11 @@ class BatchNormalizationTest(test.TestCase):
         y = math_ops.cast(x, scale.dtype) * inv + (offset - mean * inv)
         return math_ops.cast(y, x.dtype)
 
-    def _inference_ref(self, x, scale, offset, mean, var, epsilon, data_format):
+    def _inference_ref(self, x, scale, offset, mean, var, epsilon,
+                       data_format):
         if data_format not in ["NHWC", "NCHW"]:
-            raise ValueError(
-                "data_format must be NCHW or NHWC, " "got %s." % data_format
-            )
+            raise ValueError("data_format must be NCHW or NHWC, "
+                             "got %s." % data_format)
         if data_format == "NCHW":
             x = array_ops.transpose(x, [0, 2, 3, 1])
         y = self._batch_norm(x, mean, var, offset, scale, epsilon)
@@ -54,14 +54,14 @@ class BatchNormalizationTest(test.TestCase):
         return self.evaluate(y)
 
     def _test_inference(
-        self,
-        x_shape,
-        x_dtype,
-        scale_shape,
-        scale_dtype,
-        use_gpu=True,
-        exponential_avg_factor=1.0,
-        data_format="NHWC",
+            self,
+            x_shape,
+            x_dtype,
+            scale_shape,
+            scale_dtype,
+            use_gpu=True,
+            exponential_avg_factor=1.0,
+            data_format="NHWC",
     ):
         np.random.seed(1)
         x_val = np.random.random_sample(x_shape).astype(x_dtype)
@@ -89,9 +89,8 @@ class BatchNormalizationTest(test.TestCase):
                 is_training=False,
             )
             y_val = self.evaluate(y)
-            y_ref = self._inference_ref(
-                x, scale, offset, mean, var, epsilon, data_format
-            )
+            y_ref = self._inference_ref(x, scale, offset, mean, var, epsilon,
+                                        data_format)
         # An atol value of 1e-3 is too small for float16's, because some adjacent
         # float16 values that y_val can take are greater than 1e-3 apart, e.g.
         # 2.16602 and 2.16797.
@@ -105,25 +104,24 @@ class BatchNormalizationTest(test.TestCase):
             return (1.0 - factor) * old_mean + factor * new_val
 
     def _training_ref(
-        self,
-        x,
-        scale,
-        offset,
-        old_mean,
-        old_var,
-        exponential_avg_factor,
-        epsilon,
-        data_format,
+            self,
+            x,
+            scale,
+            offset,
+            old_mean,
+            old_var,
+            exponential_avg_factor,
+            epsilon,
+            data_format,
     ):
         if data_format not in ["NHWC", "NCHW"]:
-            raise ValueError(
-                "data_format must be NCHW or NHWC, " "got %s." % data_format
-            )
+            raise ValueError("data_format must be NCHW or NHWC, "
+                             "got %s." % data_format)
         if data_format == "NCHW":
             x = array_ops.transpose(x, [0, 2, 3, 1])
-        batch_mean, batch_var = nn_impl.moments(
-            math_ops.cast(x, scale.dtype), [0, 1, 2], keep_dims=False
-        )
+        batch_mean, batch_var = nn_impl.moments(math_ops.cast(x, scale.dtype),
+                                                [0, 1, 2],
+                                                keep_dims=False)
 
         y = self._batch_norm(x, batch_mean, batch_var, offset, scale, epsilon)
         if data_format == "NCHW":
@@ -133,25 +131,24 @@ class BatchNormalizationTest(test.TestCase):
         # the denominator in the formula to calculate variance, while
         # tf.compat.v1.nn.fused_batch_norm has Bessel's correction built in.
         sample_size = math_ops.cast(
-            array_ops.size(x) / array_ops.size(scale), scale.dtype
-        )
-        batch_var_corrected = (
-            batch_var * sample_size / (math_ops.maximum(sample_size - 1.0, 1.0))
-        )
+            array_ops.size(x) / array_ops.size(scale), scale.dtype)
+        batch_var_corrected = (batch_var * sample_size /
+                               (math_ops.maximum(sample_size - 1.0, 1.0)))
 
         mean = self._running_mean(old_mean, batch_mean, exponential_avg_factor)
-        var = self._running_mean(old_var, batch_var_corrected, exponential_avg_factor)
+        var = self._running_mean(old_var, batch_var_corrected,
+                                 exponential_avg_factor)
         return self.evaluate(y), self.evaluate(mean), self.evaluate(var)
 
     def _test_training(
-        self,
-        x_shape,
-        x_dtype,
-        scale_shape,
-        scale_dtype,
-        use_gpu=True,
-        exponential_avg_factor=1.0,
-        data_format="NHWC",
+            self,
+            x_shape,
+            x_dtype,
+            scale_shape,
+            scale_dtype,
+            use_gpu=True,
+            exponential_avg_factor=1.0,
+            data_format="NHWC",
     ):
         np.random.seed(1)
         x_val = np.random.random_sample(x_shape).astype(x_dtype)
@@ -161,8 +158,10 @@ class BatchNormalizationTest(test.TestCase):
             old_mean_val = None
             old_var_val = None
         else:
-            old_mean_val = np.random.random_sample(scale_shape).astype(scale_dtype)
-            old_var_val = np.random.random_sample(scale_shape).astype(scale_dtype)
+            old_mean_val = np.random.random_sample(scale_shape).astype(
+                scale_dtype)
+            old_var_val = np.random.random_sample(scale_shape).astype(
+                scale_dtype)
 
         with self.cached_session(use_gpu=use_gpu) as sess:
             x = constant_op.constant(x_val, name="x")
@@ -196,7 +195,8 @@ class BatchNormalizationTest(test.TestCase):
         self.assertAllClose(mean_ref, mean_val, atol=1e-3)
         self.assertAllClose(var_ref, var_val, atol=1e-3)
 
-    def _compute_gradient_error_float16(self, x, x32, x_shape, y, y32, y_shape):
+    def _compute_gradient_error_float16(self, x, x32, x_shape, y, y32,
+                                        y_shape):
         """Computes the gradient error for float16 inputs and/or outputs.
 
         This returns the same value as gradient_checker.compute_gradient_error. The
@@ -224,11 +224,9 @@ class BatchNormalizationTest(test.TestCase):
         # TODO(reedwm): Do not perform the unnecessary computations in
         # compute_gradient, since they double the computation time of this function.
         theoretical_grad, _ = gradient_checker.compute_gradient(
-            x, x_shape, y, y_shape, delta=1e-3, x_init_value=x_init_val
-        )
+            x, x_shape, y, y_shape, delta=1e-3, x_init_value=x_init_val)
         _, numerical_grad = gradient_checker.compute_gradient(
-            x32, x_shape, y32, y_shape, delta=1e-3, x_init_value=x32_init_val
-        )
+            x32, x_shape, y32, y_shape, delta=1e-3, x_init_value=x32_init_val)
 
         # If grad is empty, no error.
         if theoretical_grad.size == 0 and numerical_grad.size == 0:
@@ -236,15 +234,15 @@ class BatchNormalizationTest(test.TestCase):
         return np.fabs(theoretical_grad - numerical_grad).max()
 
     def _test_gradient(
-        self,
-        x_shape,
-        x_dtype,
-        scale_shape,
-        scale_dtype,
-        use_gpu=True,
-        exponential_avg_factor=1.0,
-        data_format="NHWC",
-        is_training=True,
+            self,
+            x_shape,
+            x_dtype,
+            scale_shape,
+            scale_dtype,
+            use_gpu=True,
+            exponential_avg_factor=1.0,
+            data_format="NHWC",
+            is_training=True,
     ):
         np.random.seed(1)
         x_val = np.random.random_sample(x_shape).astype(x_dtype)
@@ -259,8 +257,10 @@ class BatchNormalizationTest(test.TestCase):
                 pop_mean = None
                 pop_var = None
             else:
-                pop_mean = np.random.random_sample(scale_shape).astype(scale_dtype)
-                pop_var = np.random.random_sample(scale_shape).astype(scale_dtype)
+                pop_mean = np.random.random_sample(scale_shape).astype(
+                    scale_dtype)
+                pop_var = np.random.random_sample(scale_shape).astype(
+                    scale_dtype)
             y, _, _ = nn_impl.fused_batch_norm(
                 x,
                 scale,
@@ -272,15 +272,16 @@ class BatchNormalizationTest(test.TestCase):
                 is_training=is_training,
             )
             if x_dtype != np.float16:
-                err_x = gradient_checker.compute_gradient_error(x, x_shape, y, x_shape)
+                err_x = gradient_checker.compute_gradient_error(
+                    x, x_shape, y, x_shape)
                 err_scale = gradient_checker.compute_gradient_error(
-                    scale, scale_shape, y, x_shape
-                )
+                    scale, scale_shape, y, x_shape)
                 err_offset = gradient_checker.compute_gradient_error(
-                    offset, scale_shape, y, x_shape
-                )
+                    offset, scale_shape, y, x_shape)
             else:
-                x32 = constant_op.constant(x_val, name="x32", dtype=dtypes.float32)
+                x32 = constant_op.constant(x_val,
+                                           name="x32",
+                                           dtype=dtypes.float32)
                 y32, _, _ = nn_impl.fused_batch_norm(
                     x32,
                     scale,
@@ -292,14 +293,11 @@ class BatchNormalizationTest(test.TestCase):
                     is_training=is_training,
                 )
                 err_x = self._compute_gradient_error_float16(
-                    x, x32, x_shape, y, y32, x_shape
-                )
+                    x, x32, x_shape, y, y32, x_shape)
                 err_scale = self._compute_gradient_error_float16(
-                    scale, scale, scale_shape, y, y32, x_shape
-                )
+                    scale, scale, scale_shape, y, y32, x_shape)
                 err_offset = self._compute_gradient_error_float16(
-                    offset, offset, scale_shape, y, y32, x_shape
-                )
+                    offset, offset, scale_shape, y, y32, x_shape)
 
         x_err_tolerance = 2e-3 if x_dtype == np.float16 else 1e-3
         scale_err_tolerance = 1e-3
@@ -308,16 +306,16 @@ class BatchNormalizationTest(test.TestCase):
         self.assertLess(err_offset, scale_err_tolerance)
 
     def _test_grad_grad(
-        self,
-        x_shape,
-        x_dtype,
-        scale_shape,
-        scale_dtype,
-        use_gpu=True,
-        exponential_avg_factor=1.0,
-        data_format="NHWC",
-        is_training=True,
-        err_tolerance=1e-3,
+            self,
+            x_shape,
+            x_dtype,
+            scale_shape,
+            scale_dtype,
+            use_gpu=True,
+            exponential_avg_factor=1.0,
+            data_format="NHWC",
+            is_training=True,
+            err_tolerance=1e-3,
     ):
         np.random.seed(1)
         x_val = np.random.random_sample(x_shape).astype(x_dtype)
@@ -334,8 +332,10 @@ class BatchNormalizationTest(test.TestCase):
                 pop_mean = None
                 pop_var = None
             else:
-                pop_mean = np.random.random_sample(scale_shape).astype(scale_dtype)
-                pop_var = np.random.random_sample(scale_shape).astype(scale_dtype)
+                pop_mean = np.random.random_sample(scale_shape).astype(
+                    scale_dtype)
+                pop_var = np.random.random_sample(scale_shape).astype(
+                    scale_dtype)
             y, _, _ = nn_impl.fused_batch_norm(
                 x,
                 scale,
@@ -347,47 +347,45 @@ class BatchNormalizationTest(test.TestCase):
                 is_training=is_training,
             )
             grad_x, grad_scale, grad_offset = gradients_impl.gradients(
-                y, [x, scale, offset], grad_y
-            )
+                y, [x, scale, offset], grad_y)
 
             if is_training:
                 epsilon = y.op.get_attr("epsilon")
                 data_format = y.op.get_attr("data_format")
                 grad_vals = self.evaluate([grad_x, grad_scale, grad_offset])
-                grad_internal = nn_grad._BatchNormGrad(
-                    grad_y, x, scale, pop_mean, pop_var, epsilon, data_format
-                )
+                grad_internal = nn_grad._BatchNormGrad(grad_y, x, scale,
+                                                       pop_mean, pop_var,
+                                                       epsilon, data_format)
                 grad_internal_vals = self.evaluate(list(grad_internal))
-                for grad_val, grad_internal_val in zip(grad_vals, grad_internal_vals):
-                    self.assertAllClose(grad_val, grad_internal_val, atol=err_tolerance)
+                for grad_val, grad_internal_val in zip(grad_vals,
+                                                       grad_internal_vals):
+                    self.assertAllClose(grad_val,
+                                        grad_internal_val,
+                                        atol=err_tolerance)
 
             if x_dtype != np.float16:
                 err_grad_grad_y_1 = gradient_checker.compute_gradient_error(
-                    grad_y, x_shape, grad_x, x_shape
-                )
+                    grad_y, x_shape, grad_x, x_shape)
                 err_grad_grad_y_2 = gradient_checker.compute_gradient_error(
-                    grad_y, x_shape, grad_scale, scale_shape
-                )
+                    grad_y, x_shape, grad_scale, scale_shape)
                 err_grad_grad_y_3 = gradient_checker.compute_gradient_error(
-                    grad_y, x_shape, grad_offset, scale_shape
-                )
+                    grad_y, x_shape, grad_offset, scale_shape)
                 # In freeze mode, grad_x is not a function of x.
                 if is_training:
                     err_grad_x_1 = gradient_checker.compute_gradient_error(
-                        x, x_shape, grad_x, x_shape
-                    )
+                        x, x_shape, grad_x, x_shape)
                 err_grad_x_2 = gradient_checker.compute_gradient_error(
-                    x, x_shape, grad_scale, scale_shape
-                )
+                    x, x_shape, grad_scale, scale_shape)
 
                 err_grad_scale = gradient_checker.compute_gradient_error(
-                    scale, scale_shape, grad_x, x_shape
-                )
+                    scale, scale_shape, grad_x, x_shape)
             else:
-                x32 = constant_op.constant(x_val, dtype=dtypes.float32, name="x32")
-                grad_y32 = constant_op.constant(
-                    grad_y_val, dtype=dtypes.float32, name="grad_y32"
-                )
+                x32 = constant_op.constant(x_val,
+                                           dtype=dtypes.float32,
+                                           name="x32")
+                grad_y32 = constant_op.constant(grad_y_val,
+                                                dtype=dtypes.float32,
+                                                name="grad_y32")
                 y32, _, _ = nn_impl.fused_batch_norm(
                     x32,
                     scale,
@@ -399,29 +397,24 @@ class BatchNormalizationTest(test.TestCase):
                     is_training=is_training,
                 )
                 grad_x32, grad_scale32, grad_offset32 = gradients_impl.gradients(
-                    y32, [x32, scale, offset], grad_y32
-                )
+                    y32, [x32, scale, offset], grad_y32)
                 err_grad_grad_y_1 = self._compute_gradient_error_float16(
-                    grad_y, grad_y32, x_shape, grad_x, grad_x32, x_shape
-                )
+                    grad_y, grad_y32, x_shape, grad_x, grad_x32, x_shape)
                 err_grad_grad_y_2 = self._compute_gradient_error_float16(
-                    grad_y, grad_y32, x_shape, grad_scale, grad_scale32, scale_shape
-                )
+                    grad_y, grad_y32, x_shape, grad_scale, grad_scale32,
+                    scale_shape)
                 err_grad_grad_y_3 = self._compute_gradient_error_float16(
-                    grad_y, grad_y32, x_shape, grad_offset, grad_offset32, scale_shape
-                )
+                    grad_y, grad_y32, x_shape, grad_offset, grad_offset32,
+                    scale_shape)
                 # In freeze mode, grad_x is not a function of x.
                 if is_training:
                     err_grad_x_1 = self._compute_gradient_error_float16(
-                        x, x32, x_shape, grad_x, grad_x32, x_shape
-                    )
+                        x, x32, x_shape, grad_x, grad_x32, x_shape)
                 err_grad_x_2 = self._compute_gradient_error_float16(
-                    x, x32, x_shape, grad_scale, grad_scale32, scale_shape
-                )
+                    x, x32, x_shape, grad_scale, grad_scale32, scale_shape)
 
                 err_grad_scale = self._compute_gradient_error_float16(
-                    scale, scale, scale_shape, grad_x, grad_x32, x_shape
-                )
+                    scale, scale, scale_shape, grad_x, grad_x32, x_shape)
 
         self.assertLess(err_grad_grad_y_1, err_tolerance)
         self.assertLess(err_grad_grad_y_2, err_tolerance)
@@ -470,7 +463,8 @@ class BatchNormalizationTest(test.TestCase):
                                     np.float32,
                                     use_gpu=use_gpu,
                                     data_format=data_format,
-                                    exponential_avg_factor=exponential_avg_factor,
+                                    exponential_avg_factor=
+                                    exponential_avg_factor,
                                 )
                             else:
                                 self._test_inference(
@@ -480,7 +474,8 @@ class BatchNormalizationTest(test.TestCase):
                                     np.float32,
                                     use_gpu=use_gpu,
                                     data_format=data_format,
-                                    exponential_avg_factor=exponential_avg_factor,
+                                    exponential_avg_factor=
+                                    exponential_avg_factor,
                                 )
 
     def testInferenceShape1(self):
